Call `end` callback even if the event was already triggered
If the `end` event is emitted after the corresponding listener was
registered, the event callback is not going to be called ever.

As a workaround, we mark `emitter.ended = true` when the `end` event is
first triggered, and then stub `emitter.on()` to check for this flag and
call the callback event long after.