Fix missed first state event after a slight delay
From Resin Device Init, we have a function that returns a promise to an
operation object, therefore there is a slight delay between initiation
the operations and the client code subscribing to its `state` event,
causing in some cases to lose the first `state` event.

The solution is adding a slight delay between returning the operations
emitter object and starting to execute the operations.