Add `length` property to readable stream in burn command
Resin Image Write relies on a `length` property being present in the
readable stream that's passed to it to correctly report progress
information.

Currently, the `length` property was missing in Resin Device Operations,
losing most of progress information (eta, etc).

The fix is to use `fs.stat` to get the size of the image, and append it
as `length` is that property doesn't already exist.