Prevent duplicates from missing options
Consider an operations array containing commands that match a single
option with different values. For example:

	operations: [
		command: 'run-script'
		script: 'flashall.bat'
		when:
			os: 'win32'
	,
		command: 'run-script'
		script: 'flashall.sh'
		when:
			os: 'osx'
	,
		command: 'run-script'
		script: 'flashall.sh'
		when:
			os: 'linux'
	]

In this case, these three commands require an `os` option, and match
different values of such option.

We calculate missing options by getting all the values from `when`
properties. Therefore, in the above example, the missing options will
be:

	[ 'os', 'os', 'os' ]

This leads to issues when displaying the missing options, since the user
will see:

	Missing options: os, os and os

The solution is to simply remove the duplicates with `_.uniq`.