Implement support for "zip images"
Zip images are Zip archives that contain a single `*.iso` or `*.img`
file (they can contain any other file type as well). If this is the
case, the single image file is extracted from the Zip as a readable
stream and passed to `resin-image-write`.

If the Zip image is not valid (eg: it contains none, or multiple image
files in it) an error is thrown.